requirements.txt
# ── Core data stack ──────────────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
duckdb==0.10.3

# ── Machine learning ─────────────────────────────────────────────────────────
scikit-learn==1.5.0
lightgbm==4.3.0
xgboost==2.0.3
joblib==1.4.2

# ── NBA data ingestion ────────────────────────────────────────────────────────
nba_api==1.4.1

# ── OCR ───────────────────────────────────────────────────────────────────────
pytesseract==0.3.10
Pillow==10.3.0          # image loading for pytesseract

# ── Config / env ─────────────────────────────────────────────────────────────
python-dotenv==1.0.1
PyYAML==6.0.1

# ── HTTP / web (used in ingestion fallbacks) ──────────────────────────────────
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2

# ── CLI prettiness ────────────────────────────────────────────────────────────
rich==13.7.1
tabulate==0.9.0
click==8.1.7

# ── Utilities ────────────────────────────────────────────────────────────────
python-dateutil==2.9.0.post0
tqdm==4.66.4
